:recycle: 가게 간략 정보 바텀시트 Content 영역을 컴포넌트화 (CustomStoreInfoBottomSheetContent)
